postgres_to_es/requirements.txt
psycopg2-binary==2.9.6
python-dotenv==1.0
pydantic==1.10.7
elasticsearch==8.7.0